requirements.txt
django==2.0.10
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
